+++ demo/js/toolkit.js
@@ -13,7 +13,7 @@
     }();
     var Toolkit = {
         version: "1.4.1",
-        build: "hvsd55um",
+        build: "hvtzogp4",
         vendor: "",
         aria: true,
         messages: {
@@ -1896,7 +1896,7 @@
         this.version = "1.5.0";
         this.element = element = $(element).addClass(vendor + "matrix");
         this.options = options = this.setOptions(options, element);
-        this.items = element.find("> li");
+        this.items = [];
         this.matrix = [];
         this.wrapperWidth = 0;
         this.colWidth = 0;
@@ -1909,7 +1909,7 @@
         if (options.defer) {
             this._deferRender();
         } else {
-            this.render();
+            this.refresh();
         }
     }, {
         append: function(item) {
@@ -1925,7 +1925,10 @@
             this.refresh();
         },
         refresh: function() {
-            this.items = this.element.find("> li");
+            this.items = this.element.find("> li").each(function() {
+                var self = $(this);
+                self.addData("matrix-column-width", self.outerWidth());
+            });
             this.render();
         },
         remove: function(item) {
@@ -1941,13 +1944,14 @@
         },
         render: function() {
             this._calculateColumns();
-            if (this.items.length < this.colCount) {
-                this.element.removeAttr("style");
+            var element = this.element, items = this.items;
+            if (!items.length) {
+                element.removeAttr("style");
             } else if (this.colCount <= 1) {
-                this.element.addClass("no-columns");
-                this.items.removeAttr("style");
+                element.addClass("no-columns");
+                items.removeAttr("style");
             } else {
-                this.element.removeClass("no-columns");
+                element.removeClass("no-columns");
                 this._organizeItems();
                 this._positionItems();
             }
@@ -1978,22 +1982,24 @@
                 image.src = src;
                 promises.push(def.promise());
             });
-            $.when.apply($, promises).always(this.render);
+            $.when.apply($, promises).always(this.refresh);
         },
         _organizeItems: function() {
-            var item, span, size, c = 0, l = this.items.length;
+            var item, span, size, l = this.items.length;
             this.matrix = [];
             for (var i = 0; i < l; i++) {
                 item = this.items.eq(i);
-                size = item.outerWidth();
+                size = item.data("matrix-column-width");
                 span = Math.max(Math.round(size / this.colWidth), 1);
+                if (span > this.colCount) {
+                    span = this.colCount;
+                }
                 this.matrix.push({
                     item: item,
                     span: span
                 });
                 if (span > 1) {
                     for (var s = 1; s < span; s++) {
-                        c++;
                         if (this.matrix) {
                             this.matrix.push({
                                 item: item,
@@ -2002,14 +2008,10 @@
                         }
                     }
                 }
-                c++;
-                if (c >= this.colCount) {
-                    c = 0;
-                }
             }
         },
         _positionItems: function() {
-            var gutter = this.options.gutter, items = this.matrix, item, span, dir = this.options.rtl ? "right" : "left", x = 0, y = [], top, c = 0, i, l, s, pos = {
+            var gutter = this.options.gutter, items = this.matrix, item, span, dir = this.options.rtl ? "right" : "left", x = 0, y = [], c = 0, i, l, s, top, pos = {
                 margin: 0,
                 position: "absolute"
             };