+++ js/jquery/components/Matrix.js
@@ -11,7 +11,7 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
     this.options = options = this.setOptions(options, element);
 
     // Items within the matrix
-    this.items = element.find('> li');
+    this.items = [];
 
     // List of items in order and how many columns they span horizontally
     this.matrix = [];
@@ -40,7 +40,7 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
     if (options.defer) {
         this._deferRender();
     } else {
-        this.render();
+        this.refresh();
     }
 }, {
 
@@ -82,7 +82,13 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
      * Fetch new items and re-render the grid.
      */
     refresh: function() {
-        this.items = this.element.find('> li');
+        this.items = this.element.find('> li').each(function() {
+            var self = $(this);
+
+            // Cache the initial column width
+            self.addData('matrix-column-width', self.outerWidth());
+        });
+
         this.render();
     },
 
@@ -112,18 +118,21 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
     render: function() {
         this._calculateColumns();
 
-        // Single row, do not render
-        if (this.items.length < this.colCount) {
-            this.element.removeAttr('style');
+        var element = this.element,
+            items = this.items;
+
+        // No items
+        if (!items.length) {
+            element.removeAttr('style');
 
         // Single column
         } else if (this.colCount <= 1) {
-            this.element.addClass('no-columns');
-            this.items.removeAttr('style');
+            element.addClass('no-columns');
+            items.removeAttr('style');
 
         // Multi column
         } else {
-            this.element.removeClass('no-columns');
+            element.removeClass('no-columns');
 
             this._organizeItems();
             this._positionItems();
@@ -184,7 +193,7 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
             promises.push(def.promise());
         });
 
-        $.when.apply($, promises).always(this.render);
+        $.when.apply($, promises).always(this.refresh);
     },
 
     /**
@@ -197,17 +206,21 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
         var item,
             span,
             size,
-            c = 0,
             l = this.items.length;
 
         this.matrix = [];
 
         for (var i = 0; i < l; i++) {
             item = this.items.eq(i);
-            size = item.outerWidth();
+            size = item.data('matrix-column-width');
 
             // How many columns does this item span?
             span = Math.max(Math.round(size / this.colWidth), 1);
+
+            // Span cannot be larger than the total number of columns
+            if (span > this.colCount) {
+                span = this.colCount;
+            }
 
             this.matrix.push({
                 item: item,
@@ -217,8 +230,6 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
             // Multiple columns
             if (span > 1) {
                 for (var s = 1; s < span; s++) {
-                    c++;
-
                     if (this.matrix) {
                         this.matrix.push({
                             item: item,
@@ -226,12 +237,6 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
                         });
                     }
                 }
-            }
-
-            c++;
-
-            if (c >= this.colCount) {
-                c = 0;
             }
         }
     },
@@ -247,8 +252,13 @@ Toolkit.Matrix = Toolkit.Component.extend(function(element, options) {
             item,
             span,
             dir = this.options.rtl ? 'right' : 'left',
-            x = 0, y = [], top,
-            c = 0, i, l, s,
+            x = 0, // The left or right position value
+            y = [], // The top position values indexed by column
+            c = 0, // Current column in the loop
+            i, // Items loop counter
+            l, // Items length
+            s, // Current span column in the loop
+            top,
             pos = { margin: 0, position: 'absolute' };
 
         for (i = 0; i < this.colCount; i++) {